Changed datatype for quantity of mealIngredient to allow decimals

+++ backend/db/migrations/20230820062427-create-meal-ingredient.js
@@ -28,7 +28,7 @@ module.exports = {
         onDelete: "CASCADE"
       },
       quantity: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DECIMAL(6, 2),
         allowNull: false
       },
       unit: {
